third_party/WebKit/LayoutTests/fast/speech/scripted/start-exception-expected.txt

Test getting an exception when calling start() in the Speech JavaScript API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitSpeechRecognition' in self is true
PASS webkitSpeechRecognition == null is false

doubleStart():
window.r = new webkitSpeechRecognition()
PASS r.start() did not throw exception.
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onstart
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onaudiostart
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onsoundstart
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onspeechstart
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onnomatch
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onspeechend
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onsoundend
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onaudioend
PASS r.start() threw exception InvalidStateError: Failed to execute 'start' on 'SpeechRecognition': recognition has already started..
onend
PASS r.start() did not throw exception.
r.abort()
onend
PASS successfullyParsed is true

TEST COMPLETE
